Feature/time benchmark (#29)
* Benchmark.

* Fourier added.

+++ TDE/GPS_GRZ_class.cpp
@@ -15,7 +15,7 @@
 using namespace tde::gps;
 using namespace logger;
 
-GPS_GRZ::GPS_GRZ(uint16_t _size, uint16_t _rate, WEIGHTING_FN_TYPE _w_func) :
+GPS_GRZ::GPS_GRZ(size_t _size, size_t _rate, WEIGHTING_FN_TYPE _w_func) :
     iTDE(_size, _rate, _w_func),
     forward(_size)
 {
@@ -106,7 +106,7 @@ void GPS_GRZ::Conclude()
     unwrap_phase_spectrum(cross_phase_spectrum);
 
     //std::cout << "HARMONICAS\tCROSS_PHASE\n";
-    for (uint16_t i = 1; i < upperBound - lowerBound; ++i)
+    for (size_t i = 1; i < upperBound - lowerBound; ++i)
     {
         double harmonica = 2 * M_PI *i * sample_rate / size;
         switch(w_func)
